+++ spotbugs/src/main/java/edu/umd/cs/findbugs/util/MutableClasses.java
@@ -44,6 +44,10 @@ public class MutableClasses {
     private static final Set<String> KNOWN_IMMUTABLE_PACKAGES = new HashSet<>(Arrays.asList(
             "java.math", "java.time"));
 
+    private static final List<String> SETTER_LIKE_PREFIXES = Arrays.asList(
+            "set", "put", "add", "insert", "delete", "remove", "erase", "clear", "push", "pop",
+            "enqueue", "dequeue", "write", "append", "replace");
+
     public static boolean mutableSignature(String sig) {
         if (sig.charAt(0) == '[') {
             return true;
@@ -81,14 +85,18 @@ public class MutableClasses {
         }
     }
 
+    public static boolean looksLikeASetter(String methodName, String classSig, String retSig) {
+        if (classSig.equals(retSig)) {
+            return false;
+        }
+
+        return SETTER_LIKE_PREFIXES.stream().anyMatch(name -> methodName.startsWith(name));
+    }
+
     /**
      * Analytic information about a {@link JavaClass} relevant to determining its mutability properties.
      */
     private static final class ClassAnalysis {
-        private static final List<String> SETTER_LIKE_PREFIXES = Arrays.asList(
-                "set", "put", "add", "insert", "delete", "remove", "erase", "clear", "push", "pop",
-                "enqueue", "dequeue", "write", "append", "replace");
-
         /**
          * Class under analysis.
          */
@@ -137,15 +145,7 @@ public class MutableClasses {
         }
 
         private boolean looksLikeASetter(Method method) {
-            final String methodName = method.getName();
-            for (String name : SETTER_LIKE_PREFIXES) {
-                if (methodName.startsWith(name)) {
-                    // If setter-like methods returns an object of the same type then we suppose that it
-                    // is not a setter but creates a new instance instead.
-                    return !getSig().equals(method.getReturnType().getSignature());
-                }
-            }
-            return false;
+            return MutableClasses.looksLikeASetter(method.getName(), getSig(), method.getReturnType().getSignature());
         }
 
         private String getSig() {